Add docstrings and test for strings.convert_durations APIs for pylibcudf
